requirements.txt
django-cms
django-filer
cmsplugin-filer